Cleanup Telegraph
Move GarbageQueue to its own file
Move telegraph constants to only be local to telegraph
Move telegraph graphics into telegraph
