+++ globals.lua
@@ -13,57 +13,6 @@ this_frame_messages = {}
 server_queue = ServerQueue()
 
 score_mode = SCOREMODE_TA
-   
-TELEGRAPH_HEIGHT = 16
-TELEGRAPH_PADDING = 2 --vertical space between telegraph and stack
-TELEGRAPH_BLOCK_WIDTH = 24
-TELEGRAPH_ATTACK_MAX_SPEED = 8 --fastest an attack can travel toward the telegraph per frame
---The telegraph_attack_animation below refers the little loop shape attacks make before they start traveling toward the target.
-telegraph_attack_animation_speed = {
-    4,4,4,4,4,2,2,2,2,1,
-    1,1,1,.5,.5,.5,.5,1,1,1,
-    1,2,2,2,2,4,4,4,4,8}
-
---the following are angles out of 64, 0 being right, 32 being left, 16 being down, and 48 being up.
-telegraph_attack_animation_angles = {}
---[1] for attacks where the destination is right of the origin
-
-telegraph_attack_animation_angles[1] = {}
-for i=24,24+#telegraph_attack_animation_speed-1 do
-  telegraph_attack_animation_angles[1][#telegraph_attack_animation_angles[1]+1] = i%64
-end
---[-1] for attacks where the destination is left of the origin
-telegraph_attack_animation_angles[-1] = {}
-local leftward_animation_angle = 8
-while #telegraph_attack_animation_angles[-1] <= #telegraph_attack_animation_speed do
-  telegraph_attack_animation_angles[-1][#telegraph_attack_animation_angles[-1]+1] = leftward_animation_angle
-  leftward_animation_angle = leftward_animation_angle - 1
-  if leftward_animation_angle < 0 then
-    leftward_animation_angle = 64
-  end
-end
-
-telegraph_attack_animation = {}
-telegraph_attack_animation[1] = {}
-local leftward_or_rightward = {-1, 1}
-for k, animation in ipairs(leftward_or_rightward) do
-  telegraph_attack_animation[animation] = {}
-  for frame=1,#telegraph_attack_animation_speed do
-    local distance = telegraph_attack_animation_speed[frame]
-    local angle = telegraph_attack_animation_angles[animation][frame]/64
-    
-                --[[ use trigonometry to find the change in x and the change in y, given the hypotenuse (telegraph_attack_animation_speed) and the angle we should be traveling (2*math.pi*telegraph_attack_animation_angles[left_or_right][frame]/64)
-                
-                I think:              
-                change in y will be hypotenuse*sin angle
-                change in x will be hypotenuse*cos angle
-                --]]
-    
-    telegraph_attack_animation[animation][frame] = {}
-    telegraph_attack_animation[animation][frame].dx = distance * math.cos(angle*2*math.pi)
-    telegraph_attack_animation[animation][frame].dy = distance * math.sin(angle*2*math.pi)
-  end
-end
  
 GARBAGE_DELAY = 60
 GARBAGE_TRANSIT_TIME = 90
