+++ graphics.lua
@@ -982,132 +982,6 @@ function Stack.render_countdown(self)
   end
 end
 
-function Stack.render_telegraph(self)
-  local telegraph_to_render = self.telegraph
-  local senderCharacter = telegraph_to_render.sender.character
-
-  local render_x = telegraph_to_render.pos_x
-
-  -- Render if we are "currently chaining" for debug purposes
-  if config.debug_mode and telegraph_to_render.senderCurrentlyChaining then
-    draw(characters[senderCharacter].telegraph_garbage_images["attack"], render_x - 15 , telegraph_to_render.pos_y)
-  end
-
-  for frame_earned, attacks_this_frame in pairs(telegraph_to_render.attacks) do
-    local frames_since_earned = self.CLOCK - frame_earned
-      if frames_since_earned <= #card_animation then
-        --don't draw anything yet, card animation is still in progress.
-      elseif frames_since_earned >= GARBAGE_TRANSIT_TIME then
-        --Attack is done, remove.
-        telegraph_to_render.attacks[frame_earned] = nil
-      else
-        for _, attack in ipairs(attacks_this_frame) do
-          for _k, garbage_block in ipairs(attack.stuff_to_send) do
-            garbage_block.destination_x = telegraph_to_render.pos_x + TELEGRAPH_BLOCK_WIDTH * telegraph_to_render.garbage_queue:get_idx_of_garbage(unpack(garbage_block))
-            garbage_block.destination_y = garbage_block.destination_y or telegraph_to_render.pos_y - TELEGRAPH_HEIGHT - TELEGRAPH_PADDING 
-            
-            if not garbage_block.origin_x or not garbage_block.origin_y then
-              garbage_block.origin_x = (attack.origin_col-1) * 16 + telegraph_to_render.sender.pos_x
-              garbage_block.origin_y = (11-attack.origin_row) * 16 + telegraph_to_render.sender.pos_y + telegraph_to_render.sender.displacement - card_animation[#card_animation]
-              garbage_block.x = garbage_block.origin_x
-              garbage_block.y = garbage_block.origin_y
-              garbage_block.direction = garbage_block.direction or sign(garbage_block.destination_x - garbage_block.origin_x) --should give -1 for left, or 1 for right
-            end
-
-            if frames_since_earned <= #card_animation + #telegraph_attack_animation_speed then
-              --draw telegraph attack animation, little loop down and to the side of origin.
-     
-              -- We can't gaurantee every frame was rendered, so we must calculate the exact location regardless of how many frames happened.
-              -- TODO make this more performant?
-              garbage_block.x = garbage_block.origin_x
-              garbage_block.y = garbage_block.origin_y
-              for frame=1, frames_since_earned - #card_animation do
-                garbage_block.x = garbage_block.x + telegraph_attack_animation[garbage_block.direction][frame].dx
-                garbage_block.y = garbage_block.y + telegraph_attack_animation[garbage_block.direction][frame].dy
-              end
-
-              draw(characters[senderCharacter].telegraph_garbage_images["attack"], garbage_block.x, garbage_block.y)
-            else
-              --move toward destination
-              local distance_to_destination = math.sqrt(math.pow(garbage_block.x-garbage_block.destination_x,2)+math.pow(garbage_block.y-garbage_block.destination_y,2))
-              if frames_since_earned == #card_animation + #telegraph_attack_animation_speed then
-                garbage_block.speed = distance_to_destination / (GARBAGE_TRANSIT_TIME-frames_since_earned)
-              end
-  
-              if distance_to_destination <= (garbage_block.speed or TELEGRAPH_ATTACK_MAX_SPEED) then
-                --just move it to it's destination
-                garbage_block.x, garbage_block.y = garbage_block.destination_x, garbage_block.destination_y
-              else
-                garbage_block.x = garbage_block.x - ((garbage_block.speed or TELEGRAPH_ATTACK_MAX_SPEED)*(garbage_block.x-garbage_block.destination_x))/distance_to_destination
-                garbage_block.y = garbage_block.y - ((garbage_block.speed or TELEGRAPH_ATTACK_MAX_SPEED)*(garbage_block.y-garbage_block.destination_y))/distance_to_destination
-              end
-
-              draw(characters[senderCharacter].telegraph_garbage_images["attack"], garbage_block.x, garbage_block.y)
-            end
-          end
-        end
-      end
-  end
-
-  --then draw the telegraph's garbage queue, leaving an empty space until such a time as the attack arrives (earned_frame-GARBAGE_TRANSIT_TIME)
-  local g_queue_to_draw = telegraph_to_render.garbage_queue:makeCopy()
-  local current_block = g_queue_to_draw:pop()
-  local draw_x = telegraph_to_render.pos_x
-  local draw_y = telegraph_to_render.pos_y
-  local drewChain = false
-
-  while current_block do
-    --TODO: create a way to draw telegraphs from right to left
-    if self.CLOCK - current_block.frame_earned >= GARBAGE_TRANSIT_TIME then
-      if not current_block[3]--[[is_metal]] then
-        local height = math.min(current_block[2], 14)
-        draw(characters[senderCharacter].telegraph_garbage_images[height--[[height]]][current_block[1]--[[width]]], draw_x, draw_y)
-      else
-        draw(characters[senderCharacter].telegraph_garbage_images["metal"], draw_x, draw_y)
-      end
-      drewChain = drewChain or current_block[4]
-
-      -- Render the stop times above blocks for debug purposes
-      if config.debug_mode then
-        local stopperTime = nil
-
-        if current_block[4]--[[chain]] then
-          stopperTime = telegraph_to_render.stoppers.chain[telegraph_to_render.garbage_queue.chain_garbage.first]
-          if stopperTime and current_block.finalized then
-            stopperTime = stopperTime .. " F"
-          end
-        else
-          if current_block[3]--[[is_metal]] then
-            stopperTime = telegraph_to_render.stoppers.metal
-          else
-            stopperTime = telegraph_to_render.stoppers.combo[current_block[1]]
-          end
-        end
-
-        if stopperTime then
-          gprintf(stopperTime, draw_x*GFX_SCALE, (draw_y-8)*GFX_SCALE, 70, "center", nil, 1, large_font)
-        end
-      end
-
-    end
-    draw_x = draw_x + TELEGRAPH_BLOCK_WIDTH 
-    current_block = g_queue_to_draw:pop()
-  end
-  
-  if not drewChain and telegraph_to_render.garbage_queue.ghost_chain then
-    local draw_x = telegraph_to_render.pos_x
-    local draw_y = telegraph_to_render.pos_y
-    local height = math.min(telegraph_to_render.garbage_queue.ghost_chain, 14)
-    draw(characters[senderCharacter].telegraph_garbage_images[height][6], draw_x, draw_y)
-
-    -- Render a "G" for ghost
-    if config.debug_mode then
-      gprintf("G", draw_x*GFX_SCALE, (draw_y-8)*GFX_SCALE, 70, "center", nil, 1, large_font)
-    end
-  end
-
-end
-
 -- Draw the pause menu
 function draw_pause()
   if not GAME.renderDuringPause then
